+++ src/r_compliance.h
@@ -1,5 +1,4 @@
 // Content of this file is added to each source of fastText to change some behaviours
-
 #pragma once
 
 #include <stdlib.h>
@@ -11,6 +10,7 @@
 #define cerr Rcout // with cerr, no line refresh possible on R (it is an issue for learning with verbose set to 2, progress line is updated)
 #define cout Rcout
 #define main main_fastrtext // no direct call to main(), otherwise Cran complains + strange errors
+#define printInfo(a, b, c) interrupt_or_print(a, b, c)
 
 // catch the call to exit and call Rcpp::stop() when there is a fail
 void exit_fasttext(int error_code);