+++ NEWS.md
@@ -4,6 +4,7 @@
   * add tokenizer function
   * change the way sentence vector are computed (use fastText internal code to improve speed)
   * remove RcppThread due to change in FB source code (no more print from any threadr)
+  * add possibility to interrupte learning
 
 # 0.2.4